feat : 웹소켓을 이용한 업비트 시세 수신 기능작성

+++ src/main/java/com/example/coinProject/common/OrderBookResult.java
@@ -1,6 +1,7 @@
 package com.example.coinProject.common;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -11,5 +12,5 @@ public class OrderBookResult {
     @JsonProperty("ask_price")
     private Double askPrice;
     @JsonProperty("bid_price")
-    private Double bidPrice;
+    private BigDecimal bidPrice;
 }
